fix(all): 8 algorithmic cleanups across 6 test binaries
#1 test_cpu_multi: estimate_theoretical_bw() used CPU freq for DRAM BW → use
    global_system_config.dram_speed_mt_s (detected via dmidecode/lscpu/sysfs).
    Returns N/A instead of false estimate when DRAM speed unknown.
#2 test_memory_bandwidth: random_read_latency_thread timed each access
    individually (get_time_ns overhead 40-80% of measurement). Batched to
    1024 LCG-random reads per clock tick (matching measure_latency pattern).
#3 test_cache_hierarchy: multi_seq_write stride=16 elements vs read stride=1
    (6% data footprint). Changed to stride=8 (1 cache line), bandwidth calc
    uses 64 bytes per access (write-allocate).
#4 test_cpu_float: float/double test loops used double counters → per-iter
    float↔double conversion. Changed to long counters with explicit cast.
#5 test_cpu_multi: thread_alu_mul measured (×3+7)%1000000007 → dominant op is
    modulo, not multiply. Renamed to thread_alu_mod + updated sanity parser.
#6 test_cache_hierarchy: measure_pointer_chase_latency used get_time_ns() vs
    rdtsc_ns() (other latency functions use rdtsc_ns).
#7 test_cache_hierarchy: removed dead single_seq_read + single_bandwidth
    (never called; multi_seq_read/copy/write are the active paths).
#8 test_cpu_alu: volatile on local accumulator is intentional (prevents
    dead-code elimination and forces one store per ALU op — essential for
    accurate IPC measurement). Kept as-is; removed only the cosmetic idea.

Results: smoke 7/7, sanity 14/0/0/0, score 90.3 (A) 18/18

+++ src/test_cpu_float.c
@@ -24,11 +24,11 @@ static volatile double g_result_d = 0.0;
 
 /* ========== 浮点运算测试 ========== */
 
-/* 单精度加法测试 */
+/* 单精度加法测试 — use long counter to avoid double→float conversion per iter */
 static void test_float_add(void *arg) {
     (void)arg;
     float result = 0.0f;
-    for (double i = 0; i < ITERATIONS; i++) {
+    for (long i = 0; i < ITERATIONS; i++) {
         result += (float)i * 0.5f;
     }
     g_result_f = result;
@@ -38,7 +38,7 @@ static void test_float_add(void *arg) {
 static void test_float_mul(void *arg) {
     (void)arg;
     float result = 1.0f;
-    for (double i = 1; i < ITERATIONS; i++) {
+    for (long i = 1; i < ITERATIONS; i++) {
         result *= (float)i * 0.00001f;
         if (result < 0.0001f) result = 1.0f;
     }
@@ -50,7 +50,7 @@ static void test_float_div(void *arg) {
     (void)arg;
     float result = 1.0f;
     volatile float dividend = 1000000.0f;
-    for (double i = 1; i < ITERATIONS; i++) {
+    for (long i = 1; i < ITERATIONS; i++) {
         result = dividend / (float)i;
     }
     g_result_f = result;
@@ -60,7 +60,7 @@ static void test_float_div(void *arg) {
 static void test_float_sqrt(void *arg) {
     (void)arg;
     float result = 0.0f;
-    for (double i = 1; i < ITERATIONS; i++) {
+    for (long i = 1; i < ITERATIONS; i++) {
         result = sqrtf((float)i);
     }
     g_result_f = result;
@@ -70,8 +70,8 @@ static void test_float_sqrt(void *arg) {
 static void test_double_add(void *arg) {
     (void)arg;
     double result = 0.0;
-    for (double i = 0; i < ITERATIONS; i++) {
-        result += i * 0.5;
+    for (long i = 0; i < ITERATIONS; i++) {
+        result += (double)i * 0.5;
     }
     g_result_d = result;
 }
@@ -80,8 +80,8 @@ static void test_double_add(void *arg) {
 static void test_double_mul(void *arg) {
     (void)arg;
     double result = 1.0;
-    for (double i = 1; i < ITERATIONS; i++) {
-        result *= i * 0.00001;
+    for (long i = 1; i < ITERATIONS; i++) {
+        result *= (double)i * 0.00001;
         if (result < 0.0001) result = 1.0;
     }
     g_result_d = result;
@@ -92,8 +92,8 @@ static void test_double_div(void *arg) {
     (void)arg;
     double result = 1.0;
     volatile double dividend = 1000000.0;
-    for (double i = 1; i < ITERATIONS; i++) {
-        result = dividend / i;
+    for (long i = 1; i < ITERATIONS; i++) {
+        result = dividend / (double)i;
     }
     g_result_d = result;
 }
@@ -102,8 +102,8 @@ static void test_double_div(void *arg) {
 static void test_double_sqrt(void *arg) {
     (void)arg;
     double result = 0.0;
-    for (double i = 1; i < ITERATIONS; i++) {
-        result = sqrt(i);
+    for (long i = 1; i < ITERATIONS; i++) {
+        result = sqrt((double)i);
     }
     g_result_d = result;
 }
@@ -112,8 +112,8 @@ static void test_double_sqrt(void *arg) {
 static void test_double_sin(void *arg) {
     (void)arg;
     double result = 0.0;
-    for (double i = 1; i < ITERATIONS; i++) {
-        result += sin(i * 0.00001);
+    for (long i = 1; i < ITERATIONS; i++) {
+        result += sin((double)i * 0.00001);
     }
     g_result_d = result;
 }
@@ -122,8 +122,8 @@ static void test_double_sin(void *arg) {
 static void test_double_log(void *arg) {
     (void)arg;
     double result = 0.0;
-    for (double i = 1; i < ITERATIONS; i++) {
-        result += log(i);
+    for (long i = 1; i < ITERATIONS; i++) {
+        result += log((double)i);
     }
     g_result_d = result;
 }
@@ -132,7 +132,7 @@ static void test_double_log(void *arg) {
 static void test_double_pow(void *arg) {
     (void)arg;
     double result = 1.0;
-    for (double i = 1; i < ITERATIONS; i++) {
+    for (long i = 1; i < ITERATIONS; i++) {
         result = pow(result, 0.5);  /* sqrt via pow */
         if (result < 0.1) result = 2.0;
     }
@@ -143,8 +143,8 @@ static void test_double_pow(void *arg) {
 static void test_float_mixed(void *arg) {
     (void)arg;
     double result = 0.0;
-    for (double i = 1; i < ITERATIONS; i++) {
-        result += i * 0.5;
+    for (long i = 1; i < ITERATIONS; i++) {
+        result += (double)i * 0.5;
         result *= 1.00001;
         result = sqrt(result);
         if (result > 1000.0) result = 1.0;
